tencTech/requirements.txt
django==5.1.2
pillow
django-pagedown
django-crispy-forms
django-markdown-deux
crispy_bootstrap3